requirements.txt

python-dotenv==1.0.1
pandas==2.1.4
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.3.2
xgboost==2.0.3


mlflow==2.14.1
optuna==3.6.1
optuna-dashboard==0.15.0


fastapi==0.115.6
uvicorn==0.30.6
python-multipart==0.0.9
Jinja2==3.1.4
starlette==0.41.3


pymongo[srv]==4.8.0
certifi==2024.8.30


PyYAML==6.0.2


-e .
